backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openai>=1.0.0
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
